+++ source/xdynamics_object/xSpringDamperForce.cpp
@@ -127,7 +127,7 @@ void xSpringDamperForce::SetupDataFromListData(xTSDAData&d, std::string data)
 			else if (ch == "mass_particle_connection_list")
 			{
 				std::string ch;
-				unsigned int cnt = 0;
+				size_t cnt = 0;
 				fs >> nBodyConnection;
 				connection_body_info = new xSpringDamperBodyConnectionInfo[nBodyConnection];
 				list<xSpringDamperBodyConnectionData> bc_list;
